back_end/back_end/app/requirements.txt
pydantic
fastapi
aiomqtt
redis[asyncio]
asyncpg
python-dotenv
uvicorn